requirements.txt
attrs==24.1.0
Automat==22.10.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.6.1
filelock==3.15.4
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
IP2Location==8.10.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parsel==1.9.1
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rsa==4.9
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
zope.interface==6.4.post2
